Implemented GFGeometryCollection area algorithm. - Converted AreaOperation into a first class function call instead of a visitor. - Implemented GFGeometryCollection area algorithm. - Added GFGeometryCollection area tests. - Added MultiPolygon tests for validation of GFGeometryCollection area with multiple polygons.
